Refactor doi_data tag into a doi module.
