+++ elifecrossref/journal_article.py
@@ -1,8 +1,7 @@
-from xml.etree.ElementTree import Element, SubElement
-from elifearticle import utils as eautils
+from xml.etree.ElementTree import SubElement
 from elifecrossref import (
-    abstract, access_indicators, citation, collection, component, contributor,
-    dataset, dates, funding, related, resource_url, title)
+    abstract, access_indicators, citation, component, contributor,
+    dataset, dates, doi, funding, related, title)
 
 
 def set_journal_article(parent, poa_article, pub_date, crossref_config):
@@ -53,27 +52,12 @@ def set_journal_article(parent, poa_article, pub_date, crossref_config):
     set_archive_locations(journal_article_tag,
                           crossref_config.get("archive_locations"))
 
-    set_doi_data(journal_article_tag, poa_article, crossref_config)
+    doi.set_article_doi_data(journal_article_tag, poa_article, crossref_config)
 
     citation.set_citation_list(
         journal_article_tag, poa_article, relations_program_tag, crossref_config)
 
     component.set_component_list(journal_article_tag, poa_article, crossref_config)
-
-
-def set_doi_data(parent, poa_article, crossref_config):
-    doi_data_tag = SubElement(parent, 'doi_data')
-
-    doi_tag = SubElement(doi_data_tag, 'doi')
-    doi_tag.text = poa_article.doi
-
-    resource_tag = SubElement(doi_data_tag, 'resource')
-
-    resource = resource_url.generate_resource_url(
-        poa_article, poa_article, crossref_config)
-    resource_tag.text = resource
-
-    collection.set_collection(doi_data_tag, poa_article, "text-mining", crossref_config)
 
 
 def set_access_indicators(parent, poa_article, crossref_config):
